console: Refactor connect in gateway connection reactor
