+++ pkg/webui/locales/ja.json
@@ -538,6 +538,11 @@
   "console.containers.gateway-connection.gateway-connection.connectedTooltip": "このゲートウェイはゲートウェイサーバーに接続されていますが、ネットワークはまだゲートウェイからのアクティビティ（アップリンクやステータスメッセージの送信）を登録していません",
   "console.containers.gateway-connection.gateway-connection.otherClusterTooltip": "このゲートウェイは、このクラスタのメッセージを処理しない外部のゲートウェイサーバーに接続されています。そのため、このゲートウェイからのアクティビティを見ることはできません",
   "console.containers.gateway-connection.gateway-connection.messageCountTooltip": "最後の（再）接続以降、このゲートウェイの受信アップリンクと送信ダウンリンクの量です。ゲートウェイの種類によっては、頻繁に再接続するため、カウンタがリセットされることに注意してください",
+  "console.containers.gateway-connection.index.lastSeenAvailableTooltip": "",
+  "console.containers.gateway-connection.index.disconnectedTooltip": "",
+  "console.containers.gateway-connection.index.connectedTooltip": "",
+  "console.containers.gateway-connection.index.otherClusterTooltip": "",
+  "console.containers.gateway-connection.index.messageCountTooltip": "",
   "console.containers.gateway-location-form.gateway-location-form.updateLocationFromStatus": "",
   "console.containers.gateway-location-form.gateway-location-form.updateLocationFromStatusDescription": "",
   "console.containers.gateway-location-form.gateway-location-form.setGatewayLocation": "",
